mathpix/api-examples-master/requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
get==2019.4.13
idna==2.8
image-slicer==0.3.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
olefile==0.46
opencv-python==4.1.1.26
pdf2image==0.1.7
Pillow==6.1.0
post==2019.4.13
public==2019.4.13
pyparsing==2.4.2
pyPdf==1.13
PyPDF2==1.26.0
python-dateutil==2.8.0
python-mathpix==0.4
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Wand==0.5.7